Add EventStream
EventStream replaces the ArchiveStore as an abstraction
for traversing the events in the event store.

Compared to what it replaces, the EventStream abstraction is

  - much simpler to use

  - much simpler to implement

  - readily supported by any storage subsystem which can support
    monotonically increasing counters

It is also straightforward to extend to support sharding if needed;
each shard gets its own counter and its own EventStream. It would also
be possible to have a single multiplexing EventStream as a facade --
it could just use a (Map Shard Position) to keep track of the
position.